D2: Infrastructure for mocking of input for testing of reads()
Mocking is achieved by some bootleg macro symbol replacement of stdio
functions eg. fgetc(), with symbol for a mocked version. This mock
extracts characters from a pointer which is set by each TEST_CASE or
SECTION to a local string literal created in said SECTION etc.. Very
much malpractice and unsafe but it is the only simple way to mock IO in
C without messing with the stdin, stdout etc. file descriptors, which
can only be done in a platform specific way.