D2: Default sample entry value for table lookup reset for every test
Because Catch2 is C++ and thus the test programs are compiled as such,
the initialisation of a default nil Lox value to test table lookups has
to be constructed more manually and thus is put at the top of the file.
Unfortunately, this value was not being reset and was causing other
tests to fail. This has been corrected in the shared part of a TEST_CASE
so each SECTION tests as expected.